Fixing with chunks

+++ ping_store_check/lib/src/features/socket connection/application/socket_connection.dart
@@ -1,7 +1,9 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 import 'dart:typed_data';
 import 'package:flutter/material.dart';
+import 'package:ping_store_check/constants.dart';
 import 'package:ping_store_check/src/features/socket%20connection/data/socket_data.dart';
 import 'package:ping_store_check/src/generic%20components/my_dialog_log.dart';
 import 'package:shared_preferences/shared_preferences.dart';
@@ -9,6 +11,7 @@ import 'package:shared_preferences/shared_preferences.dart';
 Future<void> sendTo({
   required BuildContext context,
   required String msg,
+  bool recordsMode = false,
   Future<void> Function(Uint8List)? onData,
 }) async {
   SharedPreferences prefs = await SharedPreferences.getInstance();
@@ -25,8 +28,17 @@ Future<void> sendTo({
     Socket client = await Socket.connect(ip, int.parse(port));
     client.listen((data) async {
       await onData!(data);
-      client.destroy();
-      await client.close();
+      if(recordsMode) {
+        if(utf8.decode(data) == logs['success']) {
+          client.destroy();
+          await client.close();
+        } else {
+          client.write(logs['ping']);
+        }
+      } else {
+        client.destroy();
+        await client.close();
+      }
     });
     client.write(msg);
   } catch(e) {
